agent/pyproject.toml
[tool.poetry]
name = "sample_agent"
version = "0.1.0"
description = "Starter"
authors = ["Markus Ecker <markus.ecker@gmail.com>"]
license = "MIT"

[project]
name = "sample_agent"
version = "0.0.1"
dependencies = [
    "langchain-openai",
    "langchain-anthropic",
    "langchain",
    "openai",
    "langchain-community",
    "copilotkit",
    "uvicorn",
    "python-dotenv",
    "langchain-core",
    "langgraph-cli[inmem]",
    "langgraph>=0.4.10",
    "langchain-google-genai>=2.1.7",
    "google-auth-oauthlib>=1.2.2",
    "html2text>=2025.4.15",
    "beautifulsoup4>=4.13.4",
    "langchain-chroma>=0.2.4",
    "chromadb>=1.0.15",
    "hubspot-api-client>=12.0.0",
    "pyairtable>=3.1.1",
    "google-api-python-client>=2.176.0",
    "google-auth-httplib2>=0.2.0",
    "colorama>=0.4.6",
    "bs4>=0.0.2",
    "unstructured>=0.18.5",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain-openai = "^0.2.1"
langchain-anthropic = "^0.2.1"
langchain = "^0.3.1"
openai = "^1.51.0"
langchain-community = "^0.3.1"
copilotkit = "0.1.42"
uvicorn = "^0.31.0"
python-dotenv = "^1.0.1"
langchain-core = "^0.3.25"
langgraph-cli = {extras = ["inmem"], version = "^0.1.64"}

[tool.poetry.scripts]
demo = "sample_agent.demo:main"
